requirements.txt.txt
pandas
plotly.express
seaborn
matplotlib
streamlit
